Ignored error in StageOriginate?
If I read it correctly, we're currently nil here regardless of whether there is an error when doing originate or not. Shall we instead return the same value that we get from c.client.post(...)?

I tested by originating to unreachable destinations and the original version did not return an error, while the proposed one did. Please let me know if any more details could be helpful. Thanks!