Originate:  change key to referenceKey (#99)
Having the "key" field to Originate and StageOriginate is confusing.  It
is easy to think that you should pass in the key for the channel you
wish to create, rather than it being a completely optional "originator"
key.  Therefore, this patch changes the name of the variable to
`referenceKey` so as to disambiquate the two, as well as offering an
explanation in the comments.