Added ProtocolId to the ChannelData. For chan_sip and PJSIP this is the SIP header Call-ID value.
Language was in the ChannelData protobuf, but was missing from channel.go so applications could not retrieve the Language value passed from Asterisk.

Added AppName and AppData to the DialplanCEP. This allows applications to know the current Asterisk dialplan application and the parameters (data) passed to the application.

If ARI disconnects or is starting up, an application should retrieve a list of it's current channels.
Without retrieving a list of channels, we could have calls stuck in Asterisk Stasis application.
PJSIP and chan_sip calls will be disconnected when the other end disconnects.
However, local channels or ExternalMedia channels could become permanently stuck.
If the channels have stasis application name, we know we are in control of the call via ARI.
The AppData field allows us to retrieve a list of parameters that were passed to the stasis application.
This AppData is very useful when Asterisk dialplans are configured to pass data to the Stasis application.
The ARI Stasis application retrieves these parameters via the AppData (when AppName is Stasis).