Client ConnectWithContext now includes a buffered channel size setting
If the size is non-zero, a buffered channel for MsgConnected messages will be created.
The caller receives a reference to the buffered channel (non-nil indicates a buffered channel was created).
When the Client connects, it will send a MsgConnected message (true) on the buffered channel.
If the Client disconnects, it will send a MsgConnected message (false) on the buffered channel.
The caller is responsible for retrieving the buffered channel messages.# Please enter the commit message for your changes. Lines starting